refactor(debug_server): extract /touch + /input/text + injection seqlock to debug_server_input (refs #332)
Wave 23b · 13th per-family extract.  Pulls the synthetic touch
injection family + the seqlock-protected injection-state buffer
the LVGL touch driver reads from into a new sibling module
debug_server_input.{c,h}.

Endpoints (all bearer-auth):
  POST /touch       — synthetic touch injection (tap/press/release/long_press/swipe)
  POST /input/text  — type into a named LVGL textarea (e.g. chat input)

Seqlock infrastructure moved with the family:
  - debug_inject_state_t (4-field touch snapshot)
  - s_inject_state, s_inject_seq (the published state + seq counter)
  - inject_publish (writer)
  - tab5_debug_touch_override (PUBLIC reader — LVGL touch_read_cb polls this)

input_text family moved together:
  - input_text_args_t struct
  - resolve_input_target (target → LVGL textarea mapper, currently "chat")
  - input_text_apply_lvgl (LVGL-thread executor)
  - input_text_handler (httpd entry)

E2E verification on physical Tab5 192.168.1.90:
  - story_smoke 14/14 in 152s (camera shutter tap + nav taps + settings exercises /touch on every step)

debug_server.c LOC: 2,423 → ~2,070 (-353).
debug_server_input.c: 0 → 411.
